pyproject.toml
[tool.poetry]
name = "gather-your-deals"
version = "0.0.1"
description = "Python SDK for the GatherYourDeals Data API"
authors = ["yuewang"]
license = "MIT"
readme = "README.md"
packages = [{ include = "gather_your_deals" }]
repository = "https://github.com/yuewang199511/GatherYourDeals-SDK"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31"
click = "^8.1"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"
responses = "^0.25"
sphinx = "^7.0"
sphinx-rtd-theme = "^2.0"
ruff = "^0.5"
mypy = "^1.10"
types-requests = "^2.31"
types-pyyaml = "^6.0"
bandit = "^1.7"
pip-audit = "^2.7"

[tool.poetry.scripts]
gatherYourDeals = "gather_your_deals.cli.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.bandit]
exclude_dirs = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
